+++ functions/APIs/incomes.js
@@ -3,6 +3,7 @@ const { db } = require("../util/admin");
 
 exports.getAllIncomes = (request, response) => {
     db.collection("incomes")
+        .where('username', '==', request.user.username)
         .orderBy("date", "desc")
         .get()
         .then((data) => {
@@ -25,6 +26,28 @@ exports.getAllIncomes = (request, response) => {
         });
 };
 
+exports.getOneIncome = (request, response) => {
+    db.doc(`/incomes/${request.params.incomeId}`)
+        .get()
+        .then((doc) => {
+            if (!doc.exists) {
+                return response.status(404).json({ error: 'Income not found' });
+            }
+
+            if (doc.data().username !== request.user.username) {
+                return response.status(403).json({ error: 'Unauthorized' });
+            }
+
+            incomeData = doc.data();
+            incomeData.expenseId = doc.id;
+            return response.json(incomeData);
+        })
+        .catch((err) => {
+            console.log(err);
+            return response.status(500).json({ error: err.code });
+        });
+}
+
 exports.addIncome = (request, response) => {
     if (request.body.description.trim() === '') {
         return response.status(400).json({ description: "Must not be empty" });
@@ -35,6 +58,7 @@ exports.addIncome = (request, response) => {
     }
 
     const newIncomeItem = {
+        username: request.user.username,
         description: request.body.description,
         category: request.body.category,
         amount: request.body.amount,
@@ -63,6 +87,10 @@ exports.deleteIncome = (request, response) => {
         .then((doc) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: 'Income not found' });
+            }
+
+            if (doc.data().username !== request.user.username) {
+                return response.status(403).json({ error: 'Unauthorized' });
             }
 
             return document.delete();
